put stuff in it's own file, and error handling
