+++ routes/root.js
@@ -1,0 +1,9 @@
+module.exports = function rootRoute(app) {
+  app.get("/", (req, res) => {
+    const ret = {
+      status: "ok",
+      message: "server online,",
+    };
+    res.json(ret);
+  });
+};
